+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 ### Fixed
 - Subset with BBOX does not modify the input model anymore (#10)
 - `cityjson.load()` does not fail on a `GeometryInstance`, however it does not load it either (#19)
-- Fixes to the *glb* exporter (#20, #57, #83)
+- Fixes to the *glb* exporter (#20, #57, #83), and fixed the coordinate system
 
 
 ## [0.7.3] - 2021-12-15